Revert "[refactor]: 패키지 정리, 필요하지 않은 코드 제거, aws 설정, .gitignore 오타 수정"
