+++ src/main/java/com/momo/common/infrastructure/s3/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.momo.common.s3.config;
+package com.momo.common.infrastructure.s3.config;
 
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
