+++ src/main/java/com/momo/common/infrastructure/s3/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.momo.common.s3.exception;
+package com.momo.common.infrastructure.s3.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
