+++ src/main/resources/application.properties
@@ -2,15 +2,20 @@ spring.application.name=momo
 
 spring.jwt.secret=vmfhaadadadaddaddadltmskdlstkfkdgodyroqkfwkdbalroqkfwkdbalqwerqwerqwer
 
+spring.datasource.url=jdbc:mysql://localhost:3306/momo
+spring.datasource.username=root
+spring.datasource.password=zerobase24!
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
 # JPA? Hibernate ??
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.show-sql=true
 
-aws.s3.access-key=${ACCESS_KEY}
-aws.s3.secret-key=${SECRET_KEY}
-aws.s3.region=${REGION}
-aws.s3.bucket=${BUCKET}
+aws.access-key=your_access_key
+aws.secret-key=your_secret_key
+aws.region=your_region
+aws.s3.bucket=your_bucket_name
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
